# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PRM Product API
  version: 1.0.0
  description: OpenAPI specification for prm-product service.
servers:
  - url: http://localhost:8080
    description: Local development
paths:
  /products/create:
    post:
      tags:
        - Product
      summary: Create product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateProductResponse'
              example:
                message: product created successfully
                id: 1
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: invalid request body
        '500':
          description: Failed to create product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: failed to create product

  /products/update/{id}:
    patch:
      tags:
        - Product
      summary: Partially update product
      description: Only provided fields are updated. Omitted fields are left unchanged. `description` and `sale_price` may be sent as `null` to clear their values.
      operationId: updateProduct
      parameters:
        - in: path
          name: id
          required: true
          description: Product ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
              example:
                id: 1
                name: iPhone 15 Pro
                description: Updated description
                sale_price: 849.99
                price: 949.99
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidBody:
                  value:
                    error: invalid request body
                invalidId:
                  value:
                    error: invalid id format
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: product not found
        '500':
          description: Failed to update product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: failed to update product

components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: iPhone 15
        description:
          type: string
          nullable: true
          example: Smartphone
        sale_price:
          type: number
          format: double
          nullable: true
          example: 899.99
        price:
          type: number
          format: double
          example: 999.99
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - name
        - price

    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
          example: iPhone 15
        description:
          type: string
          nullable: true
          example: Smartphone
        sale_price:
          type: number
          format: double
          nullable: true
          example: 899.99
        price:
          type: number
          format: double
          example: 999.99
      required:
        - name
        - price

    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
          example: iPhone 15 Pro
        description:
          type: string
          nullable: true
          example: Updated description
        sale_price:
          type: number
          format: double
          nullable: true
          example: 849.99
        price:
          type: number
          format: double
          example: 949.99
      description: Send only fields that need to be updated.

    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: product updated successfully
      required:
        - message

    CreateProductResponse:
      type: object
      properties:
        message:
          type: string
          example: product created successfully
        id:
          type: integer
          format: int64
          example: 1
      required:
        - message
        - id

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: product not found
      required:
        - error